[minor BC break] Do not expose entity controller proxy from entity storage.
As it is being said, for CRUDL operations to entity storages should be
used, for non-CRUDL operations the entity controller services.
Therefore, it does not make too much sense to use the exposed entity
controller (proxy) from the storage to make API calls because it can
only make CRUDL calls.